+++ components/Login/index.jsx
@@ -1,15 +1,12 @@
 import React from 'react';
 import Head from 'next/head';
-import Stack from '@mui/material/Stack';
-import Container from '@mui/material/Container';
-import styles from '../../styles/Login.module.css'
+import styles from '../../styles/Login.module.css';
 
-//import login component(s) here
+// import login component(s) here
 import UserLogin from './UserLogin';
 
-
 function LoginPage() {
-  return(
+  return (
     <div>
       <Head>
         <title>Login</title>
@@ -18,8 +15,7 @@ function LoginPage() {
         <UserLogin />
       </div>
     </div>
-  )
-};
-
+  );
+}
 
 export default LoginPage;
