+++ styles/Home.module.css
@@ -1,11 +1,8 @@
-.container {
-  padding: 0 2rem;
-}
-
 .main {
-  min-height: 100vh;
+  /* min-height: 100vh;
   /* padding: 4rem 0; */
-  flex: 1;
+  /* flex: 1; */
+  width: 100vw;
   display: flex;
   flex-direction: column;
   justify-content: center;
