+++ styles/Nav.module.css
@@ -1,31 +1,45 @@
 .nav {
-  height: 50px;
+  height: 70px;
   padding: 10px;
-  background: #34495e;
-  border-radius: 8px;
-  box-shadow: 0 2px 3px 0 rgba(0,0,0,.1);
+  background: #EEE2DC;
+  box-shadow: 0 4px 4px rgba(0,0,0,.25);
+  color: #AC3B61;
   display: flex;
   align-items: center;
-  justify-content: flex-start;
-  opacity: calc(95%);
+  justify-content: center;
+  font-size: 18px;
+  font-weight: bold;
 }
 
 .nav ul {
   display: flex;
-  justify-content: center;
-  align-items: center;
   list-style: none;
+  gap: 3em;
+  margin: 0;
+  padding: 0;
+  position: relative;
 }
 
-.nav ul li a {
-  margin: 5px 15px;
-  color: #fff;
-  letter-spacing: 0.5px;
-  z-index: 1;
+.nav li {
+  transition: color 200ms ease;
 }
 
 .nav li:hover {
-  text-decoration: underline;
+  color: #68253b;
+}
+
+.navContainer {
+  max-width: 1200px;
+  width: 100vw;
+}
+
+.logoutBtn {
+  position: absolute;
+  right: 0;
+}
+
+.logoutBtn:hover {
+  cursor: pointer;
 }
